83: login flow animation
